Updated description and made url more specific
I added a description that gives a brief overview of this sub section's objectives, what it entails, and even provided a link to the fCC guide on user stories which the developer will be dealing with non-stop in this section. Finally, I updated the link to this github page to actually go directly here, and updated its text to suggest improving this intro instead of it being a stub.